requirements.txt
flask
flask-cors
requests
bleak
Pillow
pyobjc-framework-Vision
pyobjc-framework-Quartz
python-dotenv
